fixed viewer in launch files
The viewer didn't perform as expected when called through a launch file.
The reason was that, depending on the number of input arguments, a
different function was executed.

If one uses

    rosrun lsd_slam_viewer viewer

to call, the viewer would enter live slam. If more arguments were
supplied, the viewer would enter a 'read from file' mode and expect the
additional input argument to be the filepath.

To prevent this, a proper option parser was implemented, which checks
for the argument '--playBag'. If present it starts the file mode, if
missing it starts LiveSLAM.

Bag mode has NOT!! been tested.