changed openFabMap dependency to openCV3.2
Changed CMakeLists and FindOpenCV to look for the ROS kinetic version
of openCV instead of looking for a system's version.

There are further changes needed to the source to make it openCV3
compliant. They can be copied over from my GitHub fork of openFabMap:

https://github.com/FirefoxMetzger/openfabmap

It contains a bugfix (missing include) and an adapted samples folder.
The main 'src' works under openCV2 or openCV3. Only one file in samples
needs adjusting.